requirements.txt
asgiref==3.5.1
dj-static==0.0.6
Django==4.0.4
django-stdimage==5.3.0
gunicorn==20.1.0
Pillow==9.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
static3==0.7.0
